requirements.txt
python-binance==1.0.10
matplotlib==3.1.2
matplotlib-inline==0.1.2
PyYAML==5.3.1
requests==2.22.0
requests-unixsocket==0.2.0